+++ eth_defi/erc_4626/discovery_base.py
@@ -31,6 +31,7 @@ from eth_defi.erc_4626.core import ERC4262VaultDetection, ERC4626Feature, get_er
 from eth_defi.maseer_one.constants import MASEER_ONE_HARDCODED_LEADS
 from eth_defi.midas.constants import MIDAS_HARDCODED_LEADS
 from eth_defi.tokenised_fund.asseto.constants import ASSETO_HARDCODED_LEADS
+from eth_defi.tokenised_fund.franklin.constants import FRANKLIN_HARDCODED_LEADS
 from eth_defi.tokenised_fund.ondo.constants import ONDO_HARDCODED_LEADS
 from eth_defi.tokenised_fund.usyc.constants import USYC_HARDCODED_LEADS
 from eth_defi.vault.base import VaultSpec
@@ -51,6 +52,7 @@ DEFAULT_HARDCODED_VAULT_LEAD_SOURCES: HardcodedVaultLeadSources = (
     ("Vault Street", VAULT_STREET_HARDCODED_LEADS),
     ("Ondo", ONDO_HARDCODED_LEADS),
     ("Circle USYC", USYC_HARDCODED_LEADS),
+    ("Franklin Templeton", FRANKLIN_HARDCODED_LEADS),
 )
 
 if TYPE_CHECKING: